+++ src/components/Homepage/DeploymentModes/DiagramContainer/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import DeploymentOptionDiagram from '../DeploymentOptionDiagram';
+
+const DiagramContainer = ({ selectedTab }) => (
+    <DeploymentOptionDiagram selectedTab={selectedTab} />
+);
+
+export default DiagramContainer;
